sheet 2/answers/Q2/server-docker-file/requirements.txt
Flask==2.2.2
mysql_connector_repackaged==0.3.1
mysql-connector-python